Moved `expand_home_path` directory
